fix: user api & test 수정
