+++ src/test/java/com/yapp/betree/controller/AcceptanceTest.java
@@ -362,175 +362,4 @@ public class AcceptanceTest {
                 .andDo(print())
                 .andExpect(status().isOk());
     }
-
-    @Test
-    @DisplayName("유저 닉네임 변경 테스트")
-    void updateUserNickname() throws Exception {
-        User user = UserTest.TEST_USER;
-        userRepository.save(user);
-
-        String token = JwtTokenTest.JWT_TOKEN_TEST;
-        given(jwtTokenProvider.parseToken(token)).willReturn(claims(user.getId()));
-
-
-        mockMvc.perform(patch("/api/user")
-                        .header("Authorization", "Bearer " + token)
-                        .param("nickname", "닉네임 변경"))
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$").value(user.getId()));
-
-        User byId = userRepository.findById(user.getId()).get();
-        assertThat(byId.getNickname()).isEqualTo("닉네임 변경");
-
-    }
-
-    private Claims claims(Long userId) {
-        return new Claims() {
-            @Override
-            public String getIssuer() {
-                return null;
-            }
-
-            @Override
-            public Claims setIssuer(String iss) {
-                return null;
-            }
-
-            @Override
-            public String getSubject() {
-                return null;
-            }
-
-            @Override
-            public Claims setSubject(String sub) {
-                return null;
-            }
-
-            @Override
-            public String getAudience() {
-                return null;
-            }
-
-            @Override
-            public Claims setAudience(String aud) {
-                return null;
-            }
-
-            @Override
-            public Date getExpiration() {
-                return null;
-            }
-
-            @Override
-            public Claims setExpiration(Date exp) {
-                return null;
-            }
-
-            @Override
-            public Date getNotBefore() {
-                return null;
-            }
-
-            @Override
-            public Claims setNotBefore(Date nbf) {
-                return null;
-            }
-
-            @Override
-            public Date getIssuedAt() {
-                return null;
-            }
-
-            @Override
-            public Claims setIssuedAt(Date iat) {
-                return null;
-            }
-
-            @Override
-            public String getId() {
-                return null;
-            }
-
-            @Override
-            public Claims setId(String jti) {
-                return null;
-            }
-
-            @Override
-            public <T> T get(String claimName, Class<T> requiredType) {
-                return null;
-            }
-
-            @Override
-            public int size() {
-                return 0;
-            }
-
-            @Override
-            public boolean isEmpty() {
-                return false;
-            }
-
-            @Override
-            public boolean containsKey(Object key) {
-                return true;
-            }
-
-            @Override
-            public boolean containsValue(Object value) {
-                return false;
-            }
-
-            @Override
-            public Object get(Object key) {
-                String keyStr = (String) key;
-                if (keyStr.equals("id")) {
-                    return userId;
-                }
-                if (keyStr.equals("nickname")) {
-                    return "닉네임";
-                }
-                if (keyStr.equals("email")) {
-                    return "이메일";
-                }
-                return null;
-            }
-
-            @Override
-            public Object put(String key, Object value) {
-                return null;
-            }
-
-            @Override
-            public Object remove(Object key) {
-                return null;
-            }
-
-            @Override
-            public void putAll(Map<? extends String, ?> m) {
-
-            }
-
-            @Override
-            public void clear() {
-
-            }
-
-            @Override
-            public Set<String> keySet() {
-                return null;
-            }
-
-            @Override
-            public Collection<Object> values() {
-                return null;
-            }
-
-            @Override
-            public Set<Entry<String, Object>> entrySet() {
-                return null;
-            }
-        };
-    }
 }
