+++ src/test/java/com/yapp/betree/controller/UserAcceptanceTest.java
@@ -1,0 +1,220 @@
+package com.yapp.betree.controller;
+
+import com.yapp.betree.domain.User;
+import com.yapp.betree.domain.UserTest;
+import com.yapp.betree.repository.UserRepository;
+import com.yapp.betree.service.JwtTokenTest;
+import com.yapp.betree.service.oauth.JwtTokenProvider;
+import io.jsonwebtoken.Claims;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.WebApplicationContext;
+
+import java.util.Collection;
+import java.util.Date;
+import java.util.Map;
+import java.util.Set;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@Transactional
+public class UserAcceptanceTest {
+
+    private MockMvc mockMvc;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @MockBean
+    private JwtTokenProvider jwtTokenProvider;
+
+    @BeforeEach
+    void setUp(final WebApplicationContext webApplicationContext) {
+        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
+                .build();
+    }
+
+    @Test
+    @DisplayName("유저 닉네임 변경 테스트")
+    void updateUserNickname() throws Exception {
+
+        User user = userRepository.save(UserTest.TEST_USER);
+
+        String token = JwtTokenTest.JWT_TOKEN_TEST;
+        given(jwtTokenProvider.parseToken(token)).willReturn(claims(user.getId()));
+
+
+        mockMvc.perform(put("/api/users/nickname")
+                        .header("Authorization", "Bearer " + token)
+                        .param("nickname", "닉네임 변경"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$").value(user.getId()));
+
+        User byId = userRepository.findById(user.getId()).get();
+        assertThat(byId.getNickname()).isEqualTo("닉네임 변경");
+
+    }
+
+    private Claims claims(Long userId) {
+        return new Claims() {
+            @Override
+            public String getIssuer() {
+                return null;
+            }
+
+            @Override
+            public Claims setIssuer(String iss) {
+                return null;
+            }
+
+            @Override
+            public String getSubject() {
+                return null;
+            }
+
+            @Override
+            public Claims setSubject(String sub) {
+                return null;
+            }
+
+            @Override
+            public String getAudience() {
+                return null;
+            }
+
+            @Override
+            public Claims setAudience(String aud) {
+                return null;
+            }
+
+            @Override
+            public Date getExpiration() {
+                return null;
+            }
+
+            @Override
+            public Claims setExpiration(Date exp) {
+                return null;
+            }
+
+            @Override
+            public Date getNotBefore() {
+                return null;
+            }
+
+            @Override
+            public Claims setNotBefore(Date nbf) {
+                return null;
+            }
+
+            @Override
+            public Date getIssuedAt() {
+                return null;
+            }
+
+            @Override
+            public Claims setIssuedAt(Date iat) {
+                return null;
+            }
+
+            @Override
+            public String getId() {
+                return null;
+            }
+
+            @Override
+            public Claims setId(String jti) {
+                return null;
+            }
+
+            @Override
+            public <T> T get(String claimName, Class<T> requiredType) {
+                return null;
+            }
+
+            @Override
+            public int size() {
+                return 0;
+            }
+
+            @Override
+            public boolean isEmpty() {
+                return false;
+            }
+
+            @Override
+            public boolean containsKey(Object key) {
+                return true;
+            }
+
+            @Override
+            public boolean containsValue(Object value) {
+                return false;
+            }
+
+            @Override
+            public Object get(Object key) {
+                String keyStr = (String) key;
+                if (keyStr.equals("id")) {
+                    return userId;
+                }
+                if (keyStr.equals("nickname")) {
+                    return "닉네임";
+                }
+                if (keyStr.equals("email")) {
+                    return "이메일";
+                }
+                return null;
+            }
+
+            @Override
+            public Object put(String key, Object value) {
+                return null;
+            }
+
+            @Override
+            public Object remove(Object key) {
+                return null;
+            }
+
+            @Override
+            public void putAll(Map<? extends String, ?> m) {
+
+            }
+
+            @Override
+            public void clear() {
+
+            }
+
+            @Override
+            public Set<String> keySet() {
+                return null;
+            }
+
+            @Override
+            public Collection<Object> values() {
+                return null;
+            }
+
+            @Override
+            public Set<Entry<String, Object>> entrySet() {
+                return null;
+            }
+        };
+    }
+}
